Fix overflow error on images >2GB
fixes: #1926
(cherry picked from commit a4e692af87314d60c55ef9a6fd05a3bf31e6659f)

+++ pulp_container/app/models.py
@@ -120,7 +120,7 @@ class Manifest(Content):
     labels = models.JSONField(default=dict)
     architecture = models.TextField(null=True)
     os = models.TextField(null=True)
-    compressed_image_size = models.IntegerField(null=True)
+    compressed_image_size = models.BigIntegerField(null=True)
 
     # DEPRECATED: this field is deprecated and will be removed in a future release.
     is_bootable = models.BooleanField(default=False)
